feat: Add support for datetime.date in _vddd2datetime method
The `_vddd2datetime` method in the `Event` class now supports converting `vDDDTypes` objects to `datetime.date` in addition to `datetime.datetime`. This allows for more flexibility when handling different types of date values. If the `src.dt` attribute is an instance of `datetime.date`, it will be returned as is. Otherwise, if it is an instance of `datetime.datetime`, it will be converted and returned. If `src` is `None`, `None` will be returned.

This change improves the robustness of the `_vddd2datetime` method and ensures that it can handle a wider range of input types.